requirements.txt
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
mock==2.0.0
mysqlclient==1.3.9
pytest==3.0.4
SQLAlchemy==1.1.4
Flask-Cors==3.0.2